fact_checker/requirements.txt
google-adk
python-dotenv
langchain-pinecone
langchain-google-vertexai
tavily-python